Only search for quiet moves if move was a capture
This is an insanely huge optimization. Previously, Athena would search all leaf nodes until they became quiet. This was mostly redundant, however, and would cause lots of unncessary searching. Now, the quietMoveSearch() function is only called when the last move was a capture, resulting in exponentially fewer nodes needing to be searched