Fix zobrist key bug
Previously, zobrist keys where not even considering the position of the pieces upon generation. This was leading to hits on the transposition table where they shouldn't have been one. Furthermore, generating a new zobrist key everytime we manually set the position of the board ensures that we will not immediately hit transposition table entries when entering in a new position and searching